Big Update
I'm terrible at committing after every change, I'm working on that...but, this looks good on the emulated mobile site. I was able to connect to the advice slip API and display the JSON data to the html. A lot of legwork completed, will update for the desktop site tomorrow.